pyproject.toml
[project]
name = "mindiff-back"
version = "0.1.0"
description = "Mindiff Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "pymysql>=1.0.3",
    "pytest>=7.4.0",
    "hvac>=2.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.0",
]

